\name{feature_list_for_Nomogram}
\alias{feature_list_for_Nomogram}
\docType{data}
\title{
feature list for Nomogram
}
\description{
This is feature list for Nomogram.
}
\usage{data("feature_list_for_Nomogram")}
\format{
  A data frame with 6 observations on the following variable.
  \describe{
    \item{\code{ID}}{a character vector}
  }
}
\details{
This is feature list for Nomogram.
}
\value{
  Returns a plot.
}
\source{
https://portal.gdc.cancer.gov/
}

\examples{
data(feature_list_for_Nomogram)
## maybe str(feature_list_for_Nomogram)
## plot(feature_list_for_Nomogram)
}
\keyword{datasets}
