\name{Key_Clin_feature_list}
\alias{Key_Clin_feature_list}
\docType{data}
\title{
Key Clin feature list
}
\description{
This is Key Clin feature list.
}
\usage{data("Key_Clin_feature_list")}
\format{
  A data frame with 8 observations on the following variable.
  \describe{
    \item{\code{ID}}{a character vector}
  }
}
\details{
This is Key Clin feature list.
}
\value{
  Returns a feature list.
}
\source{
https://portal.gdc.cancer.gov/
}
\examples{
data(Key_Clin_feature_list)
## maybe str(Key_Clin_feature_list)

}
\keyword{datasets}
