\name{Key_Clin_features_with_PI_list}
\alias{Key_Clin_features_with_PI_list}
\docType{data}
\title{
Key Clin features with PI list
}
\description{
This is Key Clin features with PI list.
}
\usage{data("Key_Clin_features_with_PI_list")}
\format{
  A data frame with 9 observations on the following variable.
  \describe{
    \item{\code{ID}}{a character vector}
  }
}
\details{
This is Key Clin features with PI list.
}
\value{
  Returns a feature list.
}
\source{
https://portal.gdc.cancer.gov/
}

\examples{
data(Key_Clin_features_with_PI_list)
## maybe str(Key_Clin_features_with_PI_list)
}
\keyword{datasets}
