\name{Key_univariate_features_list}
\alias{Key_univariate_features_list}
\docType{data}
\title{
Key univariate features list
}
\description{
This is Key univariate features list.
}
\usage{data("Key_univariate_features_list")}
\format{
  A data frame with 2391 observations on the following variable.
  \describe{
    \item{\code{ID}}{a character vector}
  }
}
\details{
This is Key univariate features list.
}
\value{
  Returns a feature list.
}
\source{
https://portal.gdc.cancer.gov/
}

\examples{
data(Key_univariate_features_list)
## maybe str(Key_univariate_features_list)
}
\keyword{datasets}
