\name{mean_median_survival_time_data}
\alias{mean_median_survival_time_data}
\docType{data}
\title{
mean median survival time data
}
\description{
This is mean median survival time data.
}
\usage{data("mean_median_survival_time_data")}
\format{
  A data frame with 18 observations on the following 3 variables.
  \describe{
    \item{\code{IDs}}{a character vector}
    \item{\code{Mean}}{a numeric vector}
    \item{\code{Median}}{a numeric vector}
  }
}
\details{
This is mean median survival time data.
}
\value{
  Data whch is used for plot.
}
\source{
https://portal.gdc.cancer.gov/
}

\examples{
data(mean_median_survival_time_data)
# maybe str(mean_median_survival_time_data)
# plot(mean_median_survival_time_data)
}
\keyword{datasets}
