\name{survCurves_data}
\alias{survCurves_data}
\docType{data}
\title{
survCurves data
}
\description{
This is survCurves data.
}
\usage{data("survCurves_data")}
\format{
  A data frame with 15 observations on the following 19 variables.
  \describe{
    \item{\code{time_point}}{a numeric vector}
    \item{\samp{TCGA-E1-A7Z6-01}}{a numeric vector}
    \item{\samp{TCGA-S9-A7IX-01}}{a numeric vector}
    \item{\samp{TCGA-HT-8010-01}}{a numeric vector}
    \item{\samp{TCGA-VM-A8C8-01}}{a numeric vector}
    \item{\samp{TCGA-DU-5847-01}}{a numeric vector}
    \item{\samp{TCGA-TQ-A7RQ-01}}{a numeric vector}
    \item{\samp{TCGA-HT-7606-01}}{a numeric vector}
    \item{\samp{TCGA-S9-A7QY-01}}{a numeric vector}
    \item{\samp{TCGA-DH-5142-01}}{a numeric vector}
    \item{\samp{TCGA-DU-6408-01}}{a numeric vector}
    \item{\samp{TCGA-FG-8191-01}}{a numeric vector}
    \item{\samp{TCGA-TM-A84S-01}}{a numeric vector}
    \item{\samp{TCGA-DB-A64S-01}}{a numeric vector}
    \item{\samp{TCGA-DU-6542-01}}{a numeric vector}
    \item{\samp{TCGA-DU-A7T6-01}}{a numeric vector}
    \item{\samp{TCGA-DB-A4XF-01}}{a numeric vector}
    \item{\samp{TCGA-S9-A6WN-01}}{a numeric vector}
    \item{\samp{TCGA-P5-A5EX-01}}{a numeric vector}
  }
}
\details{
This is survCurves data.
}
\value{
  This is a data frame.
}
\source{
https://portal.gdc.cancer.gov/
}

\examples{
data(survCurves_data)
## maybe str(survCurves_data)
}
\keyword{datasets}
