\name{Train_PI_data}
\alias{Train_PI_data}
\docType{data}
\title{
Train PI data
}
\description{
This is Train PI data.
}
\usage{data("Train_PI_data")}
\format{
  A data frame with 158 observations on the following 57 variables.
  \describe{
    \item{\code{OS}}{a numeric vector}
    \item{\code{OS_month}}{a numeric vector}
    \item{\code{AADACL4}}{a numeric vector}
    \item{\code{ABCA12}}{a numeric vector}
    \item{\code{ABCC3}}{a numeric vector}
    \item{\code{ABI1}}{a numeric vector}
    \item{\code{ABRA}}{a numeric vector}
    \item{\code{AC006059.2}}{a numeric vector}
    \item{\code{AC008676.3}}{a numeric vector}
    \item{\code{AC008764.4}}{a numeric vector}
    \item{\code{AC010132.3}}{a numeric vector}
    \item{\code{AC012651.1}}{a numeric vector}
    \item{\code{AC013477.1}}{a numeric vector}
    \item{\code{AC109583.1}}{a numeric vector}
    \item{\code{AC113348.2}}{a numeric vector}
    \item{\code{AC117457.1}}{a numeric vector}
    \item{\code{AC131160.1}}{a numeric vector}
    \item{\code{AC144573.1}}{a numeric vector}
    \item{\code{AC235565.2}}{a numeric vector}
    \item{\code{ACO2}}{a numeric vector}
    \item{\code{ACTC1}}{a numeric vector}
    \item{\code{ACTL9}}{a numeric vector}
    \item{\code{ACTRT2}}{a numeric vector}
    \item{\code{ADD3}}{a numeric vector}
    \item{\code{ADGRL3}}{a numeric vector}
    \item{\code{ADH7}}{a numeric vector}
    \item{\code{ADPRHL1}}{a numeric vector}
    \item{\code{ADRA2A}}{a numeric vector}
    \item{\code{ADRA2B}}{a numeric vector}
    \item{\code{AFMID}}{a numeric vector}
    \item{\code{AKAP12}}{a numeric vector}
    \item{\code{AKAP14}}{a numeric vector}
    \item{\code{AKAP3}}{a numeric vector}
    \item{\code{AL139260.3}}{a numeric vector}
    \item{\code{ALG6}}{a numeric vector}
    \item{\code{ANGPTL6}}{a numeric vector}
    \item{\code{ANKRD20A1}}{a numeric vector}
    \item{\code{ANXA6}}{a numeric vector}
    \item{\code{ANXA8L1}}{a numeric vector}
    \item{\code{AOC2}}{a numeric vector}
    \item{\code{APOBEC1}}{a numeric vector}
    \item{\code{APOE}}{a numeric vector}
    \item{\code{ARHGAP11B}}{a numeric vector}
    \item{\code{ARHGAP12}}{a numeric vector}
    \item{\code{ARHGEF19}}{a numeric vector}
    \item{\code{ARL2.SNX15}}{a numeric vector}
    \item{\code{ARL6IP1}}{a numeric vector}
    \item{\code{ARPC4.TTLL3}}{a numeric vector}
    \item{\code{ASB4}}{a numeric vector}
    \item{\code{ASB5}}{a numeric vector}
    \item{\code{ASB6}}{a numeric vector}
    \item{\code{ASPM}}{a numeric vector}
    \item{\code{ATP1A2}}{a numeric vector}
    \item{\code{ATP2B4}}{a numeric vector}
    \item{\code{AZGP1}}{a numeric vector}
    \item{\code{B3GAT2}}{a numeric vector}
    \item{\code{PI}}{a numeric vector}
  }
}
\details{
This is Train PI data.
}
\value{
  A data frame with columns X and Y.
}
\source{
https://portal.gdc.cancer.gov/
}

\examples{
data(Train_PI_data)
## maybe str(Train_PI_data)
## plot(Train_PI_data)
}
\keyword{datasets}
