\name{Train_Uni_sig_data}
\alias{Train_Uni_sig_data}
\docType{data}
\title{
Train Uni sig data
}
\description{
This is Train Uni sig data.
}
\usage{data("Train_Uni_sig_data")}
\format{
  A data frame with 158 observations on the following 209 variables.
  \describe{
    \item{\code{A2ML1}}{a numeric vector}
    \item{\code{AADACL4}}{a numeric vector}
    \item{\code{AAMDC}}{a numeric vector}
    \item{\code{AAR2}}{a numeric vector}
    \item{\code{ABCA12}}{a numeric vector}
    \item{\code{ABCB4}}{a numeric vector}
    \item{\code{ABCB5}}{a numeric vector}
    \item{\code{ABCB6}}{a numeric vector}
    \item{\code{ABCC3}}{a numeric vector}
    \item{\code{ABCC9}}{a numeric vector}
    \item{\code{ABCE1}}{a numeric vector}
    \item{\code{ABCF3}}{a numeric vector}
    \item{\code{ABHD12}}{a numeric vector}
    \item{\code{ABHD14B}}{a numeric vector}
    \item{\code{ABLIM1}}{a numeric vector}
    \item{\code{ABLIM3}}{a numeric vector}
    \item{\code{ABT1}}{a numeric vector}
    \item{\code{AC004076.1}}{a numeric vector}
    \item{\code{AC005520.1}}{a numeric vector}
    \item{\code{AC005832.4}}{a numeric vector}
    \item{\code{AC006030.1}}{a numeric vector}
    \item{\code{AC008764.4}}{a numeric vector}
    \item{\code{AC010132.3}}{a numeric vector}
    \item{\code{AC013470.2}}{a numeric vector}
    \item{\code{AC027796.3}}{a numeric vector}
    \item{\code{AC097636.1}}{a numeric vector}
    \item{\code{AC106741.1}}{a numeric vector}
    \item{\code{AC131160.1}}{a numeric vector}
    \item{\code{AC132217.2}}{a numeric vector}
    \item{\code{AC144573.1}}{a numeric vector}
    \item{\code{ACAP2}}{a numeric vector}
    \item{\code{ACBD5}}{a numeric vector}
    \item{\code{ACE}}{a numeric vector}
    \item{\code{ACMSD}}{a numeric vector}
    \item{\code{ACO1}}{a numeric vector}
    \item{\code{ACO2}}{a numeric vector}
    \item{\code{ACP5}}{a numeric vector}
    \item{\code{ACRV1}}{a numeric vector}
    \item{\code{ACSBG1}}{a numeric vector}
    \item{\code{ACSM5}}{a numeric vector}
    \item{\code{ACSM6}}{a numeric vector}
    \item{\code{ACTG1}}{a numeric vector}
    \item{\code{ACTN1}}{a numeric vector}
    \item{\code{ACTR3}}{a numeric vector}
    \item{\code{ADAM12}}{a numeric vector}
    \item{\code{ADAM15}}{a numeric vector}
    \item{\code{ADAM19}}{a numeric vector}
    \item{\code{ADAM32}}{a numeric vector}
    \item{\code{ADAMTS18}}{a numeric vector}
    \item{\code{ADAMTS7}}{a numeric vector}
    \item{\code{ADAMTS8}}{a numeric vector}
    \item{\code{ADCK2}}{a numeric vector}
    \item{\code{ADCY2}}{a numeric vector}
    \item{\code{ADCY8}}{a numeric vector}
    \item{\code{ADD3}}{a numeric vector}
    \item{\code{ADGRB3}}{a numeric vector}
    \item{\code{ADGRD2}}{a numeric vector}
    \item{\code{ADGRE2}}{a numeric vector}
    \item{\code{ADGRE5}}{a numeric vector}
    \item{\code{ADGRG7}}{a numeric vector}
    \item{\code{ADGRL4}}{a numeric vector}
    \item{\code{ADGRV1}}{a numeric vector}
    \item{\code{ADH4}}{a numeric vector}
    \item{\code{ADHFE1}}{a numeric vector}
    \item{\code{ADK}}{a numeric vector}
    \item{\code{ADNP}}{a numeric vector}
    \item{\code{ADPRH}}{a numeric vector}
    \item{\code{ADRB2}}{a numeric vector}
    \item{\code{ADRM1}}{a numeric vector}
    \item{\code{AEN}}{a numeric vector}
    \item{\code{AFAP1L2}}{a numeric vector}
    \item{\code{AGAP1}}{a numeric vector}
    \item{\code{AGGF1}}{a numeric vector}
    \item{\code{AGPS}}{a numeric vector}
    \item{\code{AGR3}}{a numeric vector}
    \item{\code{AGRN}}{a numeric vector}
    \item{\code{AHCYL1}}{a numeric vector}
    \item{\code{AHNAK}}{a numeric vector}
    \item{\code{AK2}}{a numeric vector}
    \item{\code{AK8}}{a numeric vector}
    \item{\code{AKAP3}}{a numeric vector}
    \item{\code{AKAP4}}{a numeric vector}
    \item{\code{AKAP6}}{a numeric vector}
    \item{\code{AKIP1}}{a numeric vector}
    \item{\code{AKR1C1}}{a numeric vector}
    \item{\code{AL121845.3}}{a numeric vector}
    \item{\code{AL160272.2}}{a numeric vector}
    \item{\code{AL365214.2}}{a numeric vector}
    \item{\code{AL391628.1}}{a numeric vector}
    \item{\code{AL445989.1}}{a numeric vector}
    \item{\code{AL591806.2}}{a numeric vector}
    \item{\code{AL669918.1}}{a numeric vector}
    \item{\code{ALDH9A1}}{a numeric vector}
    \item{\code{ALDOB}}{a numeric vector}
    \item{\code{ALDOC}}{a numeric vector}
    \item{\code{ALG10B}}{a numeric vector}
    \item{\code{ALG2}}{a numeric vector}
    \item{\code{ALG3}}{a numeric vector}
    \item{\code{ALG6}}{a numeric vector}
    \item{\code{ALKBH3}}{a numeric vector}
    \item{\code{ALOX5AP}}{a numeric vector}
    \item{\code{ALYREF}}{a numeric vector}
    \item{\code{AMDHD1}}{a numeric vector}
    \item{\code{AMPD3}}{a numeric vector}
    \item{\code{ANGPTL2}}{a numeric vector}
    \item{\code{ANGPTL5}}{a numeric vector}
    \item{\code{ANKFY1}}{a numeric vector}
    \item{\code{ANKLE2}}{a numeric vector}
    \item{\code{ANKRD20A1}}{a numeric vector}
    \item{\code{ANKRD24}}{a numeric vector}
    \item{\code{ANKRD29}}{a numeric vector}
    \item{\code{ANKRD36}}{a numeric vector}
    \item{\code{ANKRD36C}}{a numeric vector}
    \item{\code{ANO2}}{a numeric vector}
    \item{\code{ANPEP}}{a numeric vector}
    \item{\code{ANXA1}}{a numeric vector}
    \item{\code{ANXA7}}{a numeric vector}
    \item{\code{AP003108.2}}{a numeric vector}
    \item{\code{AP3S1}}{a numeric vector}
    \item{\code{AP3S2}}{a numeric vector}
    \item{\code{APEH}}{a numeric vector}
    \item{\code{APEX2}}{a numeric vector}
    \item{\code{APH1A}}{a numeric vector}
    \item{\code{APLN}}{a numeric vector}
    \item{\code{APOBEC3B}}{a numeric vector}
    \item{\code{APOBEC3G}}{a numeric vector}
    \item{\code{APOBR}}{a numeric vector}
    \item{\code{APOE}}{a numeric vector}
    \item{\code{APOL1}}{a numeric vector}
    \item{\code{APOL4}}{a numeric vector}
    \item{\code{APPBP2}}{a numeric vector}
    \item{\code{AQP10}}{a numeric vector}
    \item{\code{AQP7}}{a numeric vector}
    \item{\code{ARAP3}}{a numeric vector}
    \item{\code{ARF4}}{a numeric vector}
    \item{\code{ARFGAP2}}{a numeric vector}
    \item{\code{ARHGAP11A}}{a numeric vector}
    \item{\code{ARHGAP11B}}{a numeric vector}
    \item{\code{ARHGAP24}}{a numeric vector}
    \item{\code{ARHGAP32}}{a numeric vector}
    \item{\code{ARHGDIG}}{a numeric vector}
    \item{\code{ARHGEF17}}{a numeric vector}
    \item{\code{ARHGEF19}}{a numeric vector}
    \item{\code{ARL1}}{a numeric vector}
    \item{\code{ARL15}}{a numeric vector}
    \item{\code{ARL2.SNX15}}{a numeric vector}
    \item{\code{ARL5C}}{a numeric vector}
    \item{\code{ARL6IP6}}{a numeric vector}
    \item{\code{ARMC4}}{a numeric vector}
    \item{\code{ARMC6}}{a numeric vector}
    \item{\code{ARMC9}}{a numeric vector}
    \item{\code{ARPC3}}{a numeric vector}
    \item{\code{ARPC4}}{a numeric vector}
    \item{\code{ARPP21}}{a numeric vector}
    \item{\code{ARSA}}{a numeric vector}
    \item{\code{ARSI}}{a numeric vector}
    \item{\code{ARSL}}{a numeric vector}
    \item{\code{ART4}}{a numeric vector}
    \item{\code{ASAH2}}{a numeric vector}
    \item{\code{ASAP1}}{a numeric vector}
    \item{\code{ASB12}}{a numeric vector}
    \item{\code{ASCC3}}{a numeric vector}
    \item{\code{ASF1B}}{a numeric vector}
    \item{\code{ASIC4}}{a numeric vector}
    \item{\code{ASPH}}{a numeric vector}
    \item{\code{ASPM}}{a numeric vector}
    \item{\code{ATAD2}}{a numeric vector}
    \item{\code{ATAD5}}{a numeric vector}
    \item{\code{ATG9B}}{a numeric vector}
    \item{\code{ATIC}}{a numeric vector}
    \item{\code{ATP11A}}{a numeric vector}
    \item{\code{ATP11C}}{a numeric vector}
    \item{\code{ATP13A4}}{a numeric vector}
    \item{\code{ATP1B2}}{a numeric vector}
    \item{\code{ATP1B3}}{a numeric vector}
    \item{\code{ATP2A3}}{a numeric vector}
    \item{\code{ATP2B1}}{a numeric vector}
    \item{\code{ATP4B}}{a numeric vector}
    \item{\code{ATP5F1D}}{a numeric vector}
    \item{\code{ATP5MC2}}{a numeric vector}
    \item{\code{ATP5ME}}{a numeric vector}
    \item{\code{ATP6V0A4}}{a numeric vector}
    \item{\code{ATP9B}}{a numeric vector}
    \item{\code{ATR}}{a numeric vector}
    \item{\code{ATRAID}}{a numeric vector}
    \item{\code{ATRIP}}{a numeric vector}
    \item{\code{ATXN7}}{a numeric vector}
    \item{\code{AUNIP}}{a numeric vector}
    \item{\code{AURKA}}{a numeric vector}
    \item{\code{AVPI1}}{a numeric vector}
    \item{\code{AWAT1}}{a numeric vector}
    \item{\code{AZU1}}{a numeric vector}
    \item{\code{B3GAT2}}{a numeric vector}
    \item{\code{B3GNT7}}{a numeric vector}
    \item{\code{B3GNT8}}{a numeric vector}
    \item{\code{B4GALNT2}}{a numeric vector}
    \item{\code{B4GALT3}}{a numeric vector}
    \item{\code{B4GALT7}}{a numeric vector}
    \item{\code{B4GAT1}}{a numeric vector}
    \item{\code{BACH2}}{a numeric vector}
    \item{\code{BAG3}}{a numeric vector}
    \item{\code{BAG5}}{a numeric vector}
    \item{\code{BAHD1}}{a numeric vector}
    \item{\code{BAIAP3}}{a numeric vector}
    \item{\code{BANK1}}{a numeric vector}
    \item{\code{BASP1}}{a numeric vector}
    \item{\code{BATF2}}{a numeric vector}
    \item{\code{BAX}}{a numeric vector}
    \item{\code{BAZ1B}}{a numeric vector}
  }
}
\details{
This is Train Uni sig data.
}
\value{
  A data frame with columns X and Y.
}
\source{
https://portal.gdc.cancer.gov/
}

\examples{
data(Train_Uni_sig_data)
## maybe str(Train_Uni_sig_data)
## plot(Train_Uni_sig_data)
}
\keyword{datasets}
