% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_test_f.R
\name{tr_test_f}
\alias{tr_test_f}
\title{This function split data into training and test data as per user defined
ratio}
\usage{
tr_test_f(data, fraction)
}
\arguments{
\item{data}{:args1 - a dataframe where features in the columns and samples must
be in rows.}

\item{fraction}{:args2 - by which user want to split data for training data;
e.g. 90\% training, so fraction would be 0.9}

}

\value{
training and test data
}
\description{
This function split data into training and test data as per user defined ratio
}
\examples{
data(New_data, package = "CPSM")
tr_test_f(data=New_data, fraction=0.9)

}




