% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_test_normalization_f.R
\name{train_test_normalization_f}
\alias{train_test_normalization_f}
\title{This function first tranform FPKM data into log2-scale transformation,
followed by quantile normalization, where training data values used as target
matrix both for training and test data for quantile  normalization}
\usage{
train_test_normalization_f(train_data, test_data, col_num)
}


\arguments{
\item{train_data}{:args1 - training data (TSV) (Patients data with clinical and
gene expression, where samples are in rows and features/genes are in columns)}
\item{test_data}{:args2 - test data (TSV) (Patients data with clinical and gene
expression, where samples are in rows and features/genes are in columns)}
\item{col_num}{:args3 - column number in data at where clinical info ends}
}


\description{
This function first tranform FPKM data into log2-scale transformation, followed
by quantile normalization, where training data values used as target matrix
both for training and test data for quantile  normalization
}
\value{
  A data frame with columns X and Y.
}
\examples{
data(train_FPKM, package = "CPSM")
data(test_FPKM, package = "CPSM")
train_test_normalization_f(train_data = train_FPKM,
test_data = test_FPKM,col_num = 21)
}





