% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Univariate_sig_features_f.R
\name{Univariate_sig_features_f}
\alias{Univariate_sig_features_f}
\title{This function selects significant features (p-value <0.05)s based on
median expression values of features using Univariate Survival analysis.}
\usage{
Univariate_sig_features_f(
  train_data,
  test_data,
  col_num,
  surv_time,
  surv_event
)
}
\arguments{
\item{train_data}{:args1 - training data (Patients data with clinical and gene
expression, where samples are in rows and features/genes are in columns)}

\item{test_data}{:args2 - Test data (Patients data with clinical and gene
expression, where samples are in rows and features/genes are in columns)}

\item{col_num}{:args3 - column number in data at where clinical info ends}

\item{surv_time}{:arg4 - name of column which contain survival time (in days)
information}

\item{surv_event}{:arg5 - name of column which contain survival
eventinformation}


}
\description{
This function selects significant features (p-value <0.05)s based on median
expression values of features using Univariate Survival analysis.
}
\value{
  A data frame with columns X and Y.
}
\examples{
data(Train_Norm_data, package = "CPSM")
data(Test_Norm_data, package = "CPSM")
Univariate_sig_features_f(train_data=Train_Norm_data,
test_data=Test_Norm_data,
col_num=21,
surv_time="OS_month" ,
surv_event="OS")
}






